% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanpv.R
\name{mesd.pv}
\alias{mesd.pv}
\title{Mean PV}
\usage{
mesd.pv(data, ...)
}
\arguments{
\item{data}{nombre de la base de datos}

\item{...}{grupo de comparación}
}
\value{
medias, desviación estándar y tamaño del efecto
}
\description{
Saca la media, desviación estándar y el tamaño del efecto de una comparación de
dos medias
}
\examples{
mesd.pv(bd,Sexo)

}
